Make asset delete worker idempotent

+++ app/workers/asset_manager_delete_asset_worker.rb
@@ -3,6 +3,8 @@ class AssetManagerDeleteAssetWorker < WorkerBase
 
   def perform(legacy_url_path)
     attributes = find_asset_by(legacy_url_path)
+    return if attributes["deleted"]
+
     asset_manager.delete_asset(attributes['id'])
   end
 end
